backend/requirements.txt
django 
djangorestframework
djongo
psycopg2
django-cors-headers
Pillow